Extract remaining setup methods to Helper module
