+++ spec/display_colour_spec.rb
@@ -1,5 +1,3 @@
-require 'display_colour'
-
 RSpec.describe DisplayColour do
   display_colour = DisplayColour.new
