+++ spec/game_factory_spec.rb
@@ -8,35 +8,12 @@ require 'player_validator'
 require 'display'
 require 'yaml'
 
-def game_factory_setup
-  display_colour = DisplayColourDouble.new
-  display = Display.new(display_colour)
-  player_validator = PlayerValidator.new(display)
-  player_factory = PlayerFactory.new(player_validator, display)
-  game_validator = GameValidator.new(display)
-  game_loader = GameLoader.new
-  filename = './spec/test_data/game_factory_test.yml'
-  GameFactory.new(player_factory, game_validator, game_loader, filename)
-end
-
-def existing_game_setup
-  filename = './spec/test_data/game_factory_test.yml'
-  game_obj = FakeClassDouble.new(1, 2, 3)
-  game_details = { 'great game' => game_obj }
-  File.open(filename, 'w') { |file| file.write game_details.to_yaml }
-end
-
-def clear_file(filename)
-  File.open(filename, 'w') { |file| file.truncate(0) }
-end
-
 RSpec.describe GameFactory do
   describe 'Creating a new game:' do
-
-    game_factory = game_factory_setup
     $stdout = StringIO.new
 
     it 'creates a human vs human game' do
+      game_factory = game_factory_setup
       allow($stdin).to receive(:gets).and_return('new', 'H', 'H', '1')
 
       game = game_factory.create_game
@@ -48,6 +25,7 @@ RSpec.describe GameFactory do
     end
 
     it 'creates a computer vs computer game' do
+      game_factory = game_factory_setup
       allow($stdin).to receive(:gets).and_return('new', 'C', 'C')
 
       game = game_factory.create_game
@@ -59,6 +37,7 @@ RSpec.describe GameFactory do
     end
 
     it 'creates a computer vs human game' do
+      game_factory = game_factory_setup
       allow($stdin).to receive(:gets).and_return('new', 'C', 'H')
 
       game = game_factory.create_game
