+++ spec/game_manager_spec.rb
@@ -1,28 +1,3 @@
-require_relative './test_doubles/display_colour_double'
-require 'display'
-require 'player_factory'
-require 'player_validator'
-require 'game_manager'
-require 'game_factory'
-require 'game_validator'
-require 'game_loader'
-require 'game_saver'
-require 'controller'
-
-def manager_setup
-  display_colour = DisplayColourDouble.new
-  display = Display.new(display_colour)
-  player_validator = PlayerValidator.new(display)
-  player_factory = PlayerFactory.new(player_validator, display)
-  game_validator = GameValidator.new(display)
-  game_loader = GameLoader.new
-  game_saver = GameSaver.new
-  filename = './spec/test_data/game_factory_test.yml'
-  game_factory = GameFactory.new(player_factory, game_validator, game_loader, filename)
-  controller = Controller.new(display, game_factory, game_saver, game_validator, filename)
-  GameManager.new(controller, display)
-end
-
 RSpec.describe GameManager do
   describe 'Play again: ' do
     it 'displays the exit message when the user inputs n' do
