+++ spec/game_saver_spec.rb
@@ -1,17 +1,3 @@
-require 'game_saver'
-require_relative './test_doubles/fake_class_double'
-require 'game'
-require 'human_player'
-require 'computer_player'
-require 'board'
-require 'display'
-require 'display_colour'
-
-
-def clear_file(filename)
-  File.open(filename, 'w') { |file| file.truncate(0) }
-end
-
 RSpec.describe GameSaver do
   describe 'Saving an object' do
     it 'saves a hash, with the key "TicTacToe" and value "text"' do
