+++ spec/game_validator_spec.rb
@@ -1,29 +1,3 @@
-require_relative './test_doubles/display_colour_double'
-require_relative './test_doubles/fake_class_double'
-require 'game_validator'
-require 'game_loader'
-require 'display'
-require 'yaml'
-
-def file_setup
-  filename = './spec/test_data/test.yml'
-  game_obj = 'game_object'
-  game_details = { 'best game' => game_obj }
-  File.open(filename, 'w') { |file| file.write game_details.to_yaml }
-  filename
-end
-
-def clear_file(filename)
-  File.open(filename, 'w') { |file| file.truncate(0) }
-end
-
-def game_validator_setup
-  display_colour = DisplayColourDouble.new
-  display = Display.new(display_colour)
-  game_loader = GameLoader.new
-  GameValidator.new(display, game_loader)
-end
-
 RSpec.describe GameValidator do
   describe 'Validating if a game name exists' do
     it "returns the user's game selection if it exists" do
